# api-definitions/Repository.openapi+json.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Repository API",
        "version": "v1"
    },
    "paths": {
...
        "/repository/demos/{demoId}": {
            "get": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "demoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "demoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/demos": {
            "get": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "gameTypes",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filterString",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "skipEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "takeEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/DemoOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Demos"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/demos/{demoId}/file": {
            "post": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "demoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DemoOrder": {
                "enum": [
                    "GameTypeAsc",
                    "GameTypeDesc",
                    "NameAsc",
                    "NameDesc",
                    "DateAsc",
                    "DateDesc",
                    "UploadedByAsc",
                    "UploadedByDesc"
                ],
                "type": "string",
                "format": "int32"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}